before additional libraries

+++ app.py
@@ -9,7 +9,7 @@ import sys
 from flask_cors import CORS, cross_origin
 from langchain.document_loaders import UnstructuredPDFLoader, OnlinePDFLoader, UnstructuredFileLoader
 from langchain.text_splitter import RecursiveCharacterTextSplitter
-from utils import recognize, audio_file_directory, pdf_file_directory, write_to_file, should_create_index, \
+from utils import recognize, audio_directory, pdf_directory, write_to_file, should_create_index, \
     texts_directory
 from langchain.vectorstores import Chroma, Pinecone
 from langchain.embeddings.openai import OpenAIEmbeddings
@@ -40,8 +40,8 @@ conversation_with_summary = ConversationChain(
 )
 
 pinecone.init(
-    api_key=pinecone_key,  # find at app.pinecone.io
-    environment=pinecone_api_env  # next to api key in console
+    api_key=pinecone_key,
+    environment=pinecone_api_env
 )
 embeddings = OpenAIEmbeddings()
 text_splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=0)
@@ -59,7 +59,6 @@ def hello_world():
 def create_message():
     message_list = request.get_json()['messages']
     message = request.get_json()['message']
-    generate_title = request.get_json()['generateTitle']
     memory = ConversationSummaryBufferMemory(llm=OpenAI())
     for elem1, elem2 in zip(message_list[::2], message_list[1::2]):
         if elem1.get('input') is not None:
@@ -69,13 +68,7 @@ def create_message():
         memory.save_context(elem1, elem2)
 
     conversation_with_summary.memory = memory
-    return jsonify({'text': conversation_with_summary.predict(input=message),
-                    'title': conversation_with_summary.predict(
-                        input="generate a title for this chat, maximum 5 words, without quotation marks")}) \
-        if generate_title else jsonify({'text': conversation_with_summary.predict(input=message)})
-    # return jsonify({'text': "raspund la intrebare",
-    #                 'title': "titlu"}) if generateTitle else \
-    #     jsonify({'text': "raspund la intrebare"})
+    return jsonify({'text': conversation_with_summary.predict(input=message)})
 
 
 @app.route('/transcribe', methods=['POST'])
@@ -86,13 +79,13 @@ def transcribe():
 
     file = request.files['file']
     filename = file.filename
-    file.save(audio_file_directory + file.filename)
+    file.save(audio_directory + file.filename)
     return recognize(filename)
 
 
 @app.route('/pdf-messages', methods=['POST'])
 @cross_origin()
-def pdf():
+def create_pdf_file():
     if 'file' not in request.files:
         return 'No file found in request!', 400
 
@@ -100,14 +93,14 @@ def pdf():
     message = request.form['message']
     file = request.files['file']
     filename = file.filename
-    file.save(pdf_file_directory + file.filename)
+    file.save(pdf_directory + file.filename)
     write_to_file(filename)
-    loader = UnstructuredFileLoader(pdf_file_directory + filename)
+    loader = UnstructuredFileLoader(pdf_directory + filename)
     data = loader.load()
     text_splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=0)
     texts = text_splitter.split_documents(data)
-
-    pinecone.delete_index(index_name)
+    if len(pinecone.list_indexes()) > 0:
+        pinecone.delete_index(index_name)
     pinecone.create_index(index_name, dimension=1536)
     docsearch = Pinecone.from_texts([t.page_content for t in texts], embeddings, index_name=index_name)
 
@@ -118,11 +111,11 @@ def pdf():
 
 @app.route('/pdf-messages/without-pdf', methods=['POST'])
 @cross_origin()
-def create_message_pdf():
+def create_pdf_message():
     # save file
     message = request.get_json()['message']
     filename = request.get_json()['filename']
-    loader = UnstructuredFileLoader(pdf_file_directory + filename)
+    loader = UnstructuredFileLoader(pdf_directory + filename)
     data = loader.load()
     texts = text_splitter.split_documents(data)
 
@@ -131,37 +124,11 @@ def create_message_pdf():
             pinecone.delete_index(index_name)
         pinecone.create_index(index_name, dimension=1536)
     docsearch = Pinecone.from_texts([t.page_content for t in texts], embeddings, index_name=index_name)
-    docs = docsearch.similarity_search(message, include_metadata=True, k=2)
+    docs = docsearch.similarity_search(message, include_metadata=True, k=4)
     text = chain.run(input_documents=docs, question=message)
 
     write_to_file(filename)
     return jsonify({'text': text})
-
-
-@app.route('/test', methods=['POST'])
-@cross_origin()
-def test():
-    write_to_file('fix')
-    x = should_create_index('fix')
-
-    return "da" if x else "nu"
-
-
-def create_index(filename):
-    pass
-
-
-def convert_bytes_to_pdf_file(bytes_string, output_path):
-    with open(output_path, 'wb') as file:
-        file.write(bytes_string)
-    print("Fișierul PDF a fost creat cu succes.")
-
-
-def write_to_file_texts(texts):
-    f = open(texts_directory + "text1.txt", 'w', encoding="utf-8")
-    for text in texts:
-        f.write(text.page_content + '\n')
-    f.close()
 
 
 if __name__ == '__main__':
